test for adding to and getting value of till

+++ app/src/main/java/codeclan/com/mariositalianrestaurant/Models/LaCarteItem.java
@@ -7,7 +7,7 @@ package codeclan.com.mariositalianrestaurant.Models;
 interface LaCarteItem {
 
     public String getName();
-    public float getPrice();
+    public double getPrice();
 
 
 //    public String getName();
